Initial commit
This repository contains everything you need to follow along with the
Minnowboard maker tutorial series up on the Minnowboard wiki provided
you have the necessary hardware for projects that require it.
The repo contains the Adafruit GPIO library, which can be found
at this repo with the purpose of creating code that can be used
across the Minnowboard MAX, Raspberry Pi, and Beaglebone Black.

Signed-off-by: Evan Steele <esteele101@gmail.com>